\name{SAE EB Logit Normal}
\alias{saelognom}
\title{SAE EB Logit Normal}
\usage{
saelognom(y, n)
}
\arguments{
  \item{y}{
  response variable contains the numbers of cases in each area.
}
  \item{n}{
  sample numbers of in each area.
}
}
\description{
This function gives the area level EB and MSE estimator to binomial model respone.
}
\examples{
\dontrun{
saelognom(y,n)
}
}
